+++ docs/docs/tutorials/codealong/contract_tutorials/advanced/token_bridge/4_typescript_glue_code.md
@@ -141,7 +141,7 @@ This fetches the wallets from the sandbox and deploys our cross chain harness on
 
 Paste the private flow test below the setup:
 
-#include_code e2e_private_cross_chain /yarn-project/end-to-end/src/e2e_cross_chain_messaging.test.ts typescript
+#include_code e2e_private_cross_chain /yarn-project/end-to-end/src/e2e_cross_chain_messaging/token_bridge_private.test.ts typescript
 
 ## Public flow test
