+++ noir-projects/noir-contracts/contracts/auth_contract/src/test.nr
