+++ yarn-project/end-to-end/src/e2e_cross_chain_messaging/token_bridge_failure_cases.test.ts
@@ -3,8 +3,8 @@ import { sha256ToField } from '@aztec/foundation/crypto';
 
 import { toFunctionSelector } from 'viem';
 
-import { CrossChainMessagingTest } from './cross_chain_messaging_test.js';
 import { NO_L1_TO_L2_MSG_ERROR } from '../fixtures/fixtures.js';
+import { CrossChainMessagingTest } from './cross_chain_messaging_test.js';
 
 describe('e2e_cross_chain_messaging token_bridge_failure_cases', () => {
   const t = new CrossChainMessagingTest('token_bridge_failure_cases');
